==> notebooks/01_Data_Wrangling.ipynb <==
# NFL Survivor Pool Data Wrangling

In this notebook we will ingest relevant datasets and clean them to prepare to use in simulations and different picking models

import numpy as np
import pandas as pd
import datetime

# Load historical odds data from
# https://www.aussportsbetting.com/data/historical-nfl-results-and-odds-data/
historical_odds = pd.read_excel("../data/nfl.xlsx")
historical_odds.head()

historical_odds.describe()

historical_odds.info()

# impute null closing lines with their opening lines since those have data for every row

We will use the closing betting lines as our odds for each game. There are missing values for these, so we will impute them with their opening lines as these are usually very different unless there is an injury another factor.

historical_odds['Home Odds Close'] = historical_odds['Home Odds Close'].fillna(historical_odds['Home Odds Open'])
historical_odds['Away Odds Close'] = historical_odds['Away Odds Close'].fillna(historical_odds['Away Odds Open'])

# remove the playoffs because most survival pools only go for the regular season
historical_odds = historical_odds[historical_odds['Playoff Game?'] != 'Y']

# create winner column
historical_odds['Winner'] = np.where(
    historical_odds['Home Score'] > historical_odds['Away Score'], historical_odds['Home Team'],
    np.where(
        historical_odds['Home Score'] < historical_odds['Away Score'], historical_odds['Away Team'],
        'tie'
    )
)

# ingest historical score from pro-football-reference.com
# Need this in order to gethe Week column
historical_scores = pd.read_csv("../data/historical_scores.csv")

historical_scores.head()

historical_scores.describe()

historical_scores.info()

historical_scores.iloc[5148:5154:, :]

# drop 'Unnamed: 0' column
historical_scores = historical_scores.drop('Unnamed: 0', axis=1)

# remove empty rows that signify the playoffs starting
historical_scores = historical_scores[historical_scores['Date'] != 'Playoffs']

# fill na values in Unnamed: 5 column with 'vs' and rename it to '@ or vs'
historical_scores['Unnamed: 5'] = historical_scores['Unnamed: 5'].fillna('vs')
historical_scores = historical_scores.rename(columns={'Unnamed: 5': '@ or vs'})

# drop Unnamed: 7
historical_scores = historical_scores.drop('Unnamed: 7', axis=1)

historical_scores['Home Team'] = np.where(
    historical_scores['@ or vs'] == 'vs', historical_scores['Winner/tie'], historical_scores['Loser/tie'])

historical_scores['Away Team'] = np.where(
    historical_scores['@ or vs'] == '@', historical_scores['Winner/tie'], historical_scores['Loser/tie'])


historical_odds['Game ID'] = historical_odds['Date'].astype(str).str[:10] + '|' + historical_odds['Home Team']
historical_scores['Game ID'] = historical_scores['Date'] + '|' + historical_scores['Home Team']

len(historical_odds)

len(pd.merge(historical_odds, historical_scores, on='Game ID', how='inner'))

merged = pd.merge(historical_odds, historical_scores, on='Game ID', how='inner')

We are losing 4 rows over all of the seasons due to date discrepencies. This is such a small amount of data loss and should be negligible. If needed we can manually enter the rows or develop a more sophisticated solution

merged.info()

# rename columns and choose ones to keep in simulation
merged = merged.rename(columns={'Date_y': 'Date',
                       'season': 'Season',
                       'Home Team_x': 'Home Team',
                       'Away Team_x': 'Away Team'
                      })
merged['Date'] = pd.to_datetime(merged['Date'])
merged = merged[['Season', 'Week', 'Date', 'Home Team', 'Away Team', 'Winner', 'Home Odds Close', 'Away Odds Close']]

# make one row for each team in a game with their odds and whether they won
new_df = pd.DataFrame(columns=['Season', 'Week', 'Date', 'Team', 'Odds', 'Won?'])
for i, row in merged.iterrows():
    home_row = {'Season': row['Season'],
                'Week': row['Week'],
                'Date': row['Date'],
                'Team': row['Home Team'],
                'Odds': row['Home Odds Close'],
                'Won?': row['Home Team'] == row['Winner']}
    away_row = {'Season': row['Season'],
                'Week': row['Week'],
                'Date': row['Date'],
                'Team': row['Away Team'],
                'Odds': row['Away Odds Close'],
                'Won?': row['Away Team'] == row['Winner']}

    new_df.loc[len(new_df)] = home_row
    new_df.loc[len(new_df)] = away_row
# add implied probability
new_df['Implied Prob'] = 1 / new_df['Odds']

cleaned_df = new_df

# this df is used as our real historical odds
cleaned_df.to_csv('../data/cleaned_odds.csv')

# this df is used to run further simulations
merged.to_csv("../data/df_for_simulation.csv")
