==> notebooks/02_Synthetic_Historical_Data_Creation.ipynb <==
# Synthetic Historical Data Creation

In this notebook I will simulate many seasons in order to have a larger dataset to test and train our picking strategies on.

import numpy as np
import pandas as pd
import random

df = pd.read_csv("../data/df_for_simulation.csv")
df.head()

def simulate_game(game_row):
    """
    Simulates the outcome of a single game based on implied win probabilities 
    derived from closing betting odds.

    Parameters:
        game_row (pd.Series): A row containing 'Home Team', 'Away Team', 
                              'Home Odds Close', and 'Away Odds Close'.

    Returns:
        str: The name of the team that is simulated to win the game.
    """
    implied_prob_home = 1/game_row['Home Odds Close']
    implied_prob_away = 1/game_row['Away Odds Close']
    implied_total = implied_prob_home + implied_prob_away
    home_prob = implied_prob_home / implied_total
    away_prob = implied_prob_away / implied_total
    teams = [game_row['Home Team'], game_row['Away Team']]
    chosen_team = np.random.choice(teams, p=[home_prob, away_prob])
    return chosen_team

simulate_game(df.iloc[0])

num_sims = 500
original_df = df.copy()
all_sims = [original_df]
for i in range(num_sims):
    simmed_history = original_df.copy()
    simmed_history['Winner'] = simmed_history.apply(simulate_game, axis=1)
    simmed_history['Season'] = simmed_history['Season'].astype(str) + '_sim_' + str(i+1)
    all_sims.append(simmed_history)
    
df = pd.concat(all_sims, ignore_index=True)

df.tail()

rows = []

for i, row in df.iterrows():
    rows.append({
        'Season': row['Season'],
        'Week': row['Week'],
        'Date': row['Date'],
        'Team': row['Home Team'],
        'Odds': row['Home Odds Close'],
        'Won?': row['Home Team'] == row['Winner']
    })
    rows.append({
        'Season': row['Season'],
        'Week': row['Week'],
        'Date': row['Date'],
        'Team': row['Away Team'],
        'Odds': row['Away Odds Close'],
        'Won?': row['Away Team'] == row['Winner']
    })

new_df = pd.DataFrame(rows)
new_df['Implied Prob'] = 1/ new_df['Odds']

new_df.to_csv("../data/simulated_nfl_histories")
