==> notebooks/03_Simulating_Upcoming_Season.ipynb <==
# Simulating Upcoming Season

In this notebook I will get odds for the upcoming season and simulate the upcoming season many times based on the odds.

import numpy as np
import random

from dotenv import load_dotenv
import os

load_dotenv()
api_key = os.getenv("MY_API_KEY")
sport = 'americanfootball_nfl'
regions = 'us'
markets = 'h2h'
oddsFormat = 'decimal'

import requests

base_url = f'https://api.the-odds-api.com/v4/sports/{sport}/odds/'

# Query parameters as a dictionary
params = {
    'apiKey': api_key,
    'regions': regions,
    'markets': markets,
    'oddsFormat': oddsFormat
}

response = requests.get(base_url, params=params)


if response.status_code == 200:
    data = response.json()  # parse JSON response
    print('Success:', data[:1])  # print first item as preview
else:
    print('Error:', response.status_code, response.text)

import pandas as pd

games = data

# Flatten
rows = []
for game in games:
    game_time = game['commence_time']
    home_team = game['home_team']
    away_team = game['away_team']

    for bookmaker in game['bookmakers']:
        bookmaker_name = bookmaker['title']
        for market in bookmaker['markets']:
            if market['key'] == 'h2h':  # focus on moneyline
                for outcome in market['outcomes']:
                    team = outcome['name']
                    odds = outcome['price']
                    rows.append({
                        'Game Start': game_time,
                        'Home Team': home_team,
                        'Away Team': away_team,
                        'Bookmaker': bookmaker_name,
                        'Team': team,
                        'Odds': odds
                    })
            

odds_df = pd.DataFrame(rows)
odds_df.head()



odds_df['Avg Odds'] = odds_df.groupby(
    ['Game Start', 'Home Team', 'Away Team', 'Team']
    )['Odds'].transform('mean')
odds_df['Game Key'] = odds_df['Game Start'] + '|' + odds_df['Home Team'] + '|' + odds_df['Away Team']
odds_df['Join Key'] = odds_df['Home Team'] + '|' + odds_df['Away Team']
avg_odds_df = odds_df.groupby(
    ['Game Start', 'Home Team', 'Away Team', 'Team', 'Game Key', 'Join Key']
    )['Odds'].mean().reset_index(name='Avg Odds')

avg_odds_df = avg_odds_df[['Game Start', 'Team', 'Avg Odds', 'Game Key', 'Join Key']].rename(columns={'Avg Odds': 'Odds'})

avg_odds_df['Implied Prob'] = 1 / avg_odds_df['Odds']
probs = []
for gk in np.unique(avg_odds_df['Game Key']):
    this_game = avg_odds_df[avg_odds_df['Game Key'] == gk] 
    p = list(this_game['Implied Prob'] / np.sum(this_game['Implied Prob']))
    probs.extend(p)
avg_odds_df['Win Prob'] = probs


avg_odds_df.head()

avg_odds_df.info()

# there are 272 regular season games in 2025, 544 is double this with 1 row for each team involved.

def simulate_game(win_probs):
    rand = np.random.rand()
    winner_idx = 0 if rand < win_probs.iloc[0] else 1
    return [True if i == winner_idx else False for i in range(len(win_probs))]

num_sims = 10000
sim_dfs = []
for i in range(num_sims):
    new_df = avg_odds_df.copy()
    new_df['Won?'] = new_df.groupby('Game Key')['Win Prob'].transform(simulate_game)
    new_df['Season'] = [i + 1 for j in range(len(new_df))]
    sim_dfs.append(new_df)

simmed_seasons = pd.concat(sim_dfs)[['Season', 'Game Start','Team', 'Odds', 'Game Key','Join Key', 'Implied Prob', 'Won?']]

simmed_seasons.head()

simmed_seasons.tail()

schedule_2025 = pd.read_csv("../data/schedule_2025.csv")
schedule_2025 = schedule_2025[schedule_2025['Week'].isin([str(week) for week in np.arange(1, 19)])]
schedule_2025.head()

schedule_2025['Join Key'] = schedule_2025['HomeTm'] + '|' + schedule_2025['VisTm']

merged = pd.merge(simmed_seasons, schedule_2025, on='Join Key')

merged[['Season', 'Week', 'Game Start','Team', 'Odds', 'Implied Prob', 'Won?']].to_csv('../data/simulated_upcoming_season')
